# Optimize Miller Loop and prepare Multi-pairing