--- a/constantine/pairing/mul_fp6_by_lines.nim
+++ b/constantine/pairing/mul_fp6_by_lines.nim
@@ -17,7 +17,8 @@ import
 # ############################################################
 #
 #                 Sparse Multiplication
-#                        by lines
+#             by lines for embedding degree 6
+#                   and sextic twist
 #
 # ############################################################