… benchmarks/bench_pairing_bls12_381.nim
… benchmarks/bench_pairing_bn254_nogami.nim
… benchmarks/bench_pairing_bn254_snarks.nim
… benchmarks/bench_pairing_template.nim
… constantine/pairing/README.md
… constantine/pairing/mul_fp12_by_lines.nim
… constantine/pairing/pairing_bls12_381.nim
… constantine/pairing/pairing_bn.nim
… constantine/primitives/bithacks.nim
… constantine/tower_field_extensions/extension_fields.nim
… docs/optimizations.md
… tests/t_pairing_mul_fp12_by_lines.nim
… tests/t_pairing_template.nim