This PR does the following:

- introduce Miller loop standard addition chains steps
- Optimize single pairing miller loop via addchain for BN254-Nogami, BLS12-377, BLS12-381
- Introduce more granularity for Fp12 * sparse elements multiplication.
  - This has a negative impact on BN curves, unlike the litterature (Scott 2019), it might not play well with the double-precision tower.
  - This gains 150 cycles over `mul(fp12, line) then mul(fp12, line)`, however except for BN254_Snarks which has a very high Hamming weight, this would occur only 5 or 6 times for a single BLS pairing and so would save 1000 cycles at most while a single pairing is about 2M cycles. Hence the complexity introduced is not worth it for single pairing. For multipairing however, this would save ~70*1000 cycles (70 being the ate loop count for BLS12 curves) while a Miller loop is 700000 cycles, so we save ~5% over douing 2 separate Miller Loops.
- Pairings API has changed to take affine points like BLST and Gurvy, so benches do not take constant-time inversion into account anymore which was artificially slowing Constantine.
- There is a file `pairing_bls12_381.nim` that will be deleted once multipairings are implementedin a generic way.